Refactor bio
